Move CBOR code into its own module, test autoAddMapDelimiters
